pyproject.toml
[tool.poetry]
name = "erever"
version = "0.1.0"
description = ""
authors = ["minaminao <minaminaoy@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.11"
pycryptodome = "^3.18.0"
# web3 = "^6.8.0"
web3 = { git = "https://github.com/minaminao/web3.py.git", branch = "develop" }


[tool.poetry.scripts]
erever = "erever.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = [
    "E501", # Line too long
]
